Migrate to stream-ordered allocator for ROCArray
