+++ src/runtime/memory/utils.jl
@@ -1,0 +1,94 @@
+function alloc_or_retry!(f)
+    for phase in 1:3
+        if phase == 2
+            GC.gc(false)
+            yield()
+        elseif phase == 3
+            GC.gc(true)
+            yield()
+        end
+
+        status = f()
+
+        @debug "Allocation phase $phase: $status"
+        if status == HSA.STATUS_SUCCESS
+            break
+        elseif status == HSA.STATUS_ERROR_OUT_OF_RESOURCES || status == HSA.STATUS_ERROR_INVALID_ALLOCATION
+            phase == 3 && check(status)
+        else
+            check(status)
+        end
+    end
+end
+
+"""
+    info()
+
+Returns a tuple of two integers, indicating respectively the free and total amount of memory
+(in bytes) available for allocation on the device.
+"""
+function info()
+    free_ref = Ref{Csize_t}()
+    total_ref = Ref{Csize_t}()
+    # FIXME: I'm not sure HSA has an API for this...
+    return convert(Int, free_ref[]), convert(Int, total_ref[])
+end
+
+"""
+    free()
+
+Returns the free amount of memory (in bytes), available for allocation on the device.
+"""
+free() = info()[1]
+
+"""
+    total()
+
+Returns the total amount of memory (in bytes), available for allocation on the device.
+"""
+total() = info()[2]
+
+"""
+    used()
+
+Returns the used amount of memory (in bytes), allocated on the device.
+"""
+used() = total()-free()
+
+"""
+Allocate linear memory on the device and return a buffer to the allocated memory. The
+allocated memory is suitably aligned for any kind of variable. The memory will not be freed
+automatically, use [`free(::Buffer)`](@ref) for that.
+"""
+function alloc end
+
+"""
+Free device memory.
+"""
+function free end
+
+"""
+Initialize device memory with a repeating value.
+"""
+function set! end
+
+"""
+Upload memory from host to device.
+Executed asynchronously on `queue` if `async` is true.
+"""
+function upload end
+@doc (@doc upload) upload!
+
+"""
+Download memory from device to host.
+Executed asynchronously on `queue` if `async` is true.
+"""
+function download end
+@doc (@doc download) download!
+
+"""
+Transfer memory from device to device.
+Executed asynchronously on `queue` if `async` is true.
+"""
+function transfer end
+@doc (@doc transfer) transfer!
