+++ test/rocarray/nmf.jl
@@ -13,9 +13,9 @@ function step(X, W, H)
     X - W * H
 end
 
-for scale in (1:5:50)
+for scale in (1:5:20)
     ncol = 2001
-    nrow = 1002*scale
+    nrow = 1002 * scale
     nfeatures = 12
 
     X = rand(Float32, nrow, ncol)
